+++ exaudfclient/base/javacontainer/script_options/BUILD
@@ -2,8 +2,8 @@ package(default_visibility = ["//visibility:public"])
 
 
 cc_library(
-    name = "java_scriptoptionlines",
+    name = "java_script_option_lines",
     hdrs = [":converter.h"],
     srcs = [":parser.h", ":converter.h", ":converter.cc", ":parser_legacy.h", ":parser_legacy.cc"],
-    deps = ["//base/script_options_parser:scriptoptionlinesparser", "//base/exaudflib:header", "//base/exaudflib:exaudflib-deps"],
+    deps = ["//base/script_options_parser:script_option_lines_parser", "//base/exaudflib:header", "//base/exaudflib:exaudflib-deps"],
 )
